pyproject.toml
[tool.poetry]
name = "ustscripts"
version = "0.1.0"
description = ""
authors = ["faith <jamienibtong@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
gcsa = "^2.3.0"
browser-cookie3 = "^0.19.1"
bs4 = "^0.0.2"
html5lib = "^1.1"
pywebview = "^4.4.1"
black = "^24.8.0"
httpx = "^0.27.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false

[tool.pyright]
exclude = [ ".venv" ]
venvPath = "."
venv = ".venv"
